+++ .idea/gradle.xml
@@ -12,6 +12,8 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/local" />
+            <option value="$PROJECT_DIR$/remote" />
           </set>
         </option>
       </GradleProjectSettings>
