+++ app/build.gradle
@@ -46,12 +46,16 @@ android {
 
 dependencies {
 
+    implementation project(path: ':local')
+    implementation project(path: ':remote')
+
     implementation 'androidx.core:core-ktx:1.10.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.8.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
     implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
+
 
     //Unit test
     testImplementation 'junit:junit:4.13.2'
@@ -61,8 +65,6 @@ dependencies {
     //Compose
     implementation 'androidx.compose.ui:ui:1.4.3'
     implementation "androidx.compose.ui:ui-tooling-preview:1.4.3"
-
-
 
     // ViewModel
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.1"
@@ -89,9 +91,6 @@ dependencies {
     // Firebase UI
     implementation 'com.firebaseui:firebase-ui-auth:8.0.2'
 
-    //retrofit
-    implementation "com.squareup.retrofit2:retrofit:2.9.0"
-    implementation "com.squareup.retrofit2:converter-gson:2.9.0"
 
     //Corrutinas
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.4'
@@ -101,18 +100,10 @@ dependencies {
     kapt("com.google.dagger:hilt-android-compiler:2.44")
 
 
-    implementation 'com.squareup.okhttp3:okhttp:4.10.0'
-    implementation 'com.squareup.okhttp3:logging-interceptor:4.10.0'
 
     //Glide
     implementation "com.github.bumptech.glide:glide:4.14.2"
     annotationProcessor "com.github.bumptech.glide:compiler:4.14.2"
-
-    //Room
-    implementation 'androidx.room:room-runtime:2.5.1'
-    implementation "androidx.room:room-ktx:2.5.1"
-    kapt "androidx.room:room-compiler:2.5.1"
-    testImplementation("androidx.room:room-testing:2.5.1")
