+++ app/src/main/java/com/edith/movies/domain/GetMoviesUseCase.kt
@@ -1,8 +1,7 @@
 package com.edith.movies.domain
 
-
-import com.edith.movies.local.model.toDatabase
 import com.edith.movies.uimodel.Movie
+import com.edith.movies.uimodel.toDatabase
 import javax.inject.Inject
 
 class GetMoviesUseCase @Inject constructor(private val repository: MoviesDbRepository) {
