+++ app/src/main/java/com/edith/movies/domain/MoviesDbRepository.kt
@@ -1,11 +1,11 @@
 package com.edith.movies.domain
 
 
-import com.edith.movies.local.model.MovieEntity
-import com.edith.movies.remote.model.GenderResponse
-import com.edith.movies.remote.model.LastMoviesResponse
-import com.edith.movies.remote.model.MovieDb
-import com.edith.movies.remote.model.MovieModel
+import com.edith.local.model.MovieEntity
+import com.edith.remote.model.GenderResponse
+import com.edith.remote.model.LastMoviesResponse
+import com.edith.remote.model.MovieDb
+import com.edith.remote.model.MovieModel
 import com.edith.movies.uimodel.Movie
 
 interface MoviesDbRepository {
