--- app/src/main/java/com/edith/movies/local/dao/GenderDao.kt
@@ -1,19 +1,0 @@
-package com.edith.movies.local.dao
-
-import androidx.room.Dao
-import androidx.room.Delete
-import androidx.room.Insert
-import androidx.room.Query
-import com.edith.movies.local.model.GenderEntity
-
-@Dao
-interface GenderDao {
-    @Query("SELECT * FROM gender")
-    fun getAll(): List<GenderEntity>
-
-    @Insert
-    fun insertAll(vararg genderEntity: GenderEntity)
-
-    @Delete
-    fun delete(genderEntity: GenderEntity)
-}
