--- app/src/main/java/com/edith/movies/remote/model/DatesModel.kt
@@ -1,6 +1,0 @@
-package com.edith.movies.remote.model
-
-data class DatesModel(
-    val maximum: String,
-    val minimum: String
-)
