--- app/src/main/java/com/edith/movies/remote/model/GenderModel.kt
@@ -1,7 +1,0 @@
-package com.edith.movies.remote.model
-
-data class GenderModel(
-    val id: Int,
-    val name: String
-)
-
