--- app/src/main/java/com/edith/movies/remote/model/GenresMoviesResponse.kt
@@ -1,11 +1,0 @@
-package com.edith.movies.remote.model
-
-data class GenresMoviesResponse(
-    val genres: List<GenderResponse>
-)
-
-data class GenderResponse(
-    val id: Int,
-    val string: String
-)
-
