--- app/src/main/java/com/edith/movies/remote/model/MoviesDbResponse.kt
@@ -1,14 +1,0 @@
-package com.edith.movies.remote.model
-
-import android.os.Parcelable
-import com.google.gson.annotations.SerializedName
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class MoviesDbResponse(
-
-    @SerializedName("page") val page: Int,
-    @SerializedName("results") val results: List<MovieDb>,
-    @SerializedName("total_pages") val total_pages: Int,
-    @SerializedName("total_results") val total_results: Int
-) : Parcelable
