+++ app/src/main/java/com/edith/movies/view/adapter/MoviesAdapter.kt
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.edith.movies.remote.model.MovieDb
+import com.edith.remote.model.MovieDb
 import com.edith.movies.databinding.ViewMovieItemBinding
