+++ remote/src/main/java/com/edith/remote/model/DatesModel.kt
@@ -1,0 +1,6 @@
+package com.edith.remote.model
+
+data class DatesModel(
+    val maximum: String,
+    val minimum: String
+)
