+++ remote/src/main/java/com/edith/remote/model/GenderModel.kt
@@ -1,0 +1,7 @@
+package com.edith.remote.model
+
+data class GenderModel(
+    val id: Int,
+    val name: String
+)
+
