+++ remote/src/main/java/com/edith/remote/model/GenresMoviesResponse.kt
@@ -1,0 +1,11 @@
+package com.edith.remote.model
+
+data class GenresMoviesResponse(
+    val genres: List<GenderResponse>
+)
+
+data class GenderResponse(
+    val id: Int,
+    val string: String
+)
+
