+++ remote/src/main/java/com/edith/remote/model/TopRatedMoviesResponse.kt
@@ -1,0 +1,8 @@
+package com.edith.remote.model
+
+data class TopRatedMoviesResponse(
+    val page: Int,
+    val results: List<MovieModel>,
+    val total_pages: Int,
+    val total_results: Int
+)
