+++ settings.gradle
@@ -14,3 +14,5 @@ dependencyResolutionManagement {
 }
 rootProject.name = "Movies"
 include ':app'
+include ':local'
+include ':remote'
